Refactored result lists into RaceListsUC.*
