+++ RaceHorology/RaceListsUC.xaml
@@ -1,0 +1,91 @@
+﻿<UserControl x:Class="RaceHorology.RaceListsUC"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:RaceHorology"
+             xmlns:racehorologylib="clr-namespace:RaceHorologyLib;assembly=RaceHorologyLib"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+  <UserControl.Resources>
+    <racehorologylib:PositionConverter x:Key="PositionConverter" />
+    <!--
+    <local:BooleanToBrushConverter x:Key="BooleanToBrushConverter" />
+    <racehorologylib:ResultCodeConverter x:Key="ResultCodeConverter" />
+    <racehorologylib:ResultCodeWithCommentConverter x:Key="ResultCodeWithCommentConverter" />
+    -->
+  </UserControl.Resources>
+
+  <Grid Grid.Column="2" >
+    <Grid.RowDefinitions>
+      <RowDefinition Height="Auto"/>
+      <RowDefinition Height="*"/>
+      <RowDefinition Height="Auto"/>
+    </Grid.RowDefinitions>
+
+    <Grid Width="Auto">
+      <Grid.ColumnDefinitions>
+        <ColumnDefinition Width="Auto"/>
+        <ColumnDefinition Width="*"/>
+        <ColumnDefinition Width="Auto"/>
+      </Grid.ColumnDefinitions>
+      <ComboBox x:Name="cmbTotalResult" Margin="5" Grid.Column="0" SelectionChanged="CmbTotalResult_SelectionChanged"/>
+      <ComboBox x:Name="cmbTotalResultGrouping" Margin="5" Grid.Column="2" SelectionChanged="CmbTotalResultGrouping_SelectionChanged"/>
+    </Grid>
+
+    <DataGrid FontSize="12" x:Name="dgTotalResults" Margin="5,0,5,0" Grid.Row="1" AlternationCount="2" AlternatingRowBackground="#FFf8faf7" AutoGenerateColumns="False" CanUserAddRows="False" CanUserSortColumns="False" EnableColumnVirtualization="True" EnableRowVirtualization="True" VirtualizingPanel.IsVirtualizingWhenGrouping="True">
+      <DataGrid.Columns>
+        <DataGridTextColumn local:DataGridUtil.Name="Position" Header="Platz" Binding="{Binding Position, Mode=OneWay, Converter={StaticResource PositionConverter}}" />
+        <DataGridTextColumn local:DataGridUtil.Name="StartNumber" Header="Startnummer" Binding="{Binding Path=Participant.StartNumber, Mode=OneWay}" />
+        <DataGridTextColumn local:DataGridUtil.Name="Name" Header="Name" Binding="{Binding Path=Participant.Name, Mode=OneWay}" />
+        <DataGridTextColumn local:DataGridUtil.Name="Firstname" Header="Vorname" Binding="{Binding Path=Participant.Firstname, Mode=OneWay}" />
+        <DataGridTextColumn local:DataGridUtil.Name="Year" Header="Jahrgang" Binding="{Binding Path=Participant.Year, Mode=OneWay}" />
+        <DataGridTextColumn local:DataGridUtil.Name="Class" Header="Klasse" Binding="{Binding Path=Participant.Class, Mode=OneWay}" />
+        <DataGridTextColumn local:DataGridUtil.Name="Club" Header="Verein" Binding="{Binding Path=Participant.Club, Mode=OneWay}" />
+      </DataGrid.Columns>
+
+      <DataGrid.Style>
+        <Style TargetType="DataGrid">
+          <Setter Property="AlternatingRowBackground" Value="#FFf8faf7"/>
+        </Style>
+      </DataGrid.Style>
+
+      <DataGrid.GroupStyle>
+        <GroupStyle>
+          <GroupStyle.HeaderTemplate>
+            <DataTemplate>
+              <TextBlock FontWeight="Bold" FontSize="14" Text="{Binding Name}"/>
+            </DataTemplate>
+          </GroupStyle.HeaderTemplate>
+        </GroupStyle>
+      </DataGrid.GroupStyle>
+
+      <DataGrid.RowStyle>
+        <Style TargetType="DataGridRow">
+          <Style.Triggers>
+            <DataTrigger Binding="{Binding JustModified}" Value="true">
+              <Setter Property="Background" Value="Orange" />
+            </DataTrigger>
+          </Style.Triggers>
+        </Style>
+      </DataGrid.RowStyle>
+
+    </DataGrid>
+
+    <Grid Width="Auto" Grid.Row="2">
+      <Grid.ColumnDefinitions>
+        <ColumnDefinition Width="*"/>
+        <ColumnDefinition Width="Auto"/>
+        <ColumnDefinition Width="Auto"/>
+        <ColumnDefinition Width="Auto"/>
+        <ColumnDefinition Width="Auto"/>
+        <ColumnDefinition Width="Auto"/>
+      </Grid.ColumnDefinitions>
+      <Button x:Name="btnExportDSVAlpin" Content="Export DSVAlpin" Margin="5" Grid.Row="0" Grid.Column="1" Click="BtnExportDsvAlpin_Click"/>
+      <Button x:Name="btnExportCSV" Content="Export CSV" Margin="5" Grid.Row="0" Grid.Column="2" Click="BtnExportCsv_Click"/>
+      <Button x:Name="btnExportXLSX" Content="Export Excel" Margin="5" Grid.Row="0" Grid.Column="3" Click="BtnExportXlsx_Click"/>
+      <Button x:Name="btnExport" Content="Export DSV" Margin="5" Grid.Row="0" Grid.Column="4" Click="BtnExportDsv_Click"/>
+      <Button x:Name="btnPrint" Content="Drucken" Margin="5" Grid.Row="0" Grid.Column="5" Click="BtnPrint_Click"/>
+    </Grid>
+  </Grid>
+</UserControl>
